working versions of everything
