+++ test/s3_data_aggregator/cdk.json
